# app/web/api/openapi/openapi.yaml
openapi: 3.0.2
info:
  description: API for "rpi-pwm" enabling users to remotely control app settings (e.g.,
    fan speeds)
  license:
    name: GNU AGPLv3
    url: https://www.gnu.org/licenses/agpl.txt
  title: rpi-pwm
  version: 1.0.0
servers:
- description: Local dev server
  url: http://127.0.0.1:{port}/api/{api-version}
  variables:
    port:
      default: "8080"
      description: Port of webserver
    api-version:
      default: v1
      description: Version of API
- description: Local reverse proxy setup
  url: https://rpi-pwm.rpi.lan/api/{api-version}
  variables:
    api-version:
      default: v1
      description: Version of API
security:
- api-security: []
paths:
  /app/config:
    description: Singleton resource of config flags
    get:
      description: Returns current config flags
      operationId: app_config_get
      responses:
        "200":
          content:
            application/json:
              examples:
                app-config_singleton:
                  value:
                    fanOn: true
                    loggingEnabled: true
                    loggingLevel: WARN
                    DCUpdateIntervalInSec: 3
                    selectedFanCurve:
                      id: 916CD0EB-A755-4663-8410-461431039F74
                      name: Quiet
                      fanCurveSeries:
                      - tempInCels: 30
                        fanDCInPerc: 40
                      - tempInCels: 35
                        fanDCInPerc: 50
                    pwmGpioPin: 12
                    pwmInvertSignal: true
                    pwmMinDCInPerc: 20
                    pwmMaxDCInPerc: 95
              schema:
                $ref: '#/components/schemas/AppConfig'
          description: Config of app
          headers:
            ETag:
              examples:
                etag-example:
                  value: b5395865858689cfcb718a8fbf84e128771f6975
              explode: false
              schema:
                type: string
              style: simple
      summary: Returns current config flags
      x-openapi-router-controller: app.web.api.controllers.default_controller
    put:
      description: Updates config flags
      operationId: app_config_put
      parameters:
      - description: Current hash of to be updated config -> used for optimistic locking
        examples:
          If-Match-example:
            value: b5395865858689cfcb718a8fbf84e128771f6975
        explode: false
        in: header
        name: If-Match
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            examples:
              app-config_singleton:
                value:
                  app:
                    fanOn: true
                    loggingEnabled: true
                    loggingLevel: WARN
                    DCUpdateIntervalInSec: 3
                    selectedFanCurve:
                      id: 916CD0EB-A755-4663-8410-461431039F74
                      name: Quiet
                      fanCurveSeries:
                      - tempInCels: 30
                        fanDCInPerc: 40
                      - tempInCels: 35
                        fanDCInPerc: 50
                  pwm:
                    gpioPin: 12
                    invertSignal: true
                    minDCInPerc: 20
                    maxDCInPerc: 95
            schema:
              $ref: '#/components/schemas/AppConfig'
        description: Config flags consisting of app and pwm config
        required: true
      responses:
        "200":
          content:
            application/json:
              examples:
                app-config_singleton:
                  value:
                    fanOn: true
                    loggingEnabled: true
                    loggingLevel: WARN
                    DCUpdateIntervalInSec: 3
                    selectedFanCurve:
                      id: 916CD0EB-A755-4663-8410-461431039F74
                      name: Quiet
                      fanCurveSeries:
                      - tempInCels: 30
                        fanDCInPerc: 40
                      - tempInCels: 35
                        fanDCInPerc: 50
                    pwmGpioPin: 12
                    pwmInvertSignal: true
                    pwmMinDCInPerc: 20
                    pwmMaxDCInPerc: 95
              schema:
                $ref: '#/components/schemas/AppConfig'
          description: Config of app
          headers:
            ETag:
              examples:
                etag-example:
                  value: b5395865858689cfcb718a8fbf84e128771f6975
              explode: false
              schema:
                type: string
              style: simple
        "400":
          description: The request failed due to a malformed request body
      summary: Updates config flags
      x-openapi-router-controller: app.web.api.controllers.default_controller
    summary: Allows user to view config of app
  /app/fanCurves:
    description: Collection resource for fan curves
    get:
      description: Returns list of all available fan curves
      operationId: app_fan_curves_get
      parameters:
      - description: Filter for fan curves whose name is similar to `name`
        examples:
          name:
            value: Quiet-
        explode: true
        in: query
        name: name
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              examples:
                app-fanCurves_collection:
                  value:
                  - id: 916CD0EB-A755-4663-8410-461431039F74
                    name: Quiet
                    fanCurveSeries:
                    - tempInCels: 30
                      fanDCInPerc: 40
                    - tempInCels: 35
                      fanDCInPerc: 50
                  - id: 316CD0EB-A755-4463-8410-461411039F74
                    name: Quiet-2
                    fanCurveSeries:
                    - tempInCels: 30
                      fanDCInPerc: 20
                    - tempInCels: 35
                      fanDCInPerc: 10
                  - id: 016ED0EB-A755-4663-8210-461431039F74
                    name: Loud
                    fanCurveSeries:
                    - tempInCels: 20
                      fanDCInPerc: 50
                    - tempInCels: 35
                      fanDCInPerc: 90
                    - tempInCels: 45
                      fanDCInPerc: 100
              schema:
                items:
                  $ref: '#/components/schemas/AppFanCurve'
                type: array
          description: Multiple fan curves
      summary: Returns list of all available fan curves
      x-openapi-router-controller: app.web.api.controllers.default_controller
    post:
      description: Adds new fan curve
      operationId: app_fan_curves_post
      parameters:
      - description: Filter for fan curves whose name is similar to `name`
        examples:
          name:
            value: Quiet-
        explode: true
        in: query
        name: name
        required: false
        schema:
          type: string
        style: form
      requestBody:
        content:
          application/json:
            examples:
              app-fanCurvesBase_singleton:
                value:
                  name: Quiet
                  fanCurveSeries:
                  - tempInCels: 30
                    fanDCInPerc: 40
                  - tempInCels: 35
                    fanDCInPerc: 50
            schema:
              $ref: '#/components/schemas/AppFanCurveBase'
        required: true
      responses:
        "201":
          content:
            application/json:
              examples:
                app-fanCurves_singleton:
                  value:
                    id: 916CD0EB-A755-4663-8410-461431039F74
                    name: Quiet
                    fanCurveSeries:
                    - tempInCels: 30
                      fanDCInPerc: 40
                    - tempInCels: 35
                      fanDCInPerc: 50
              schema:
                $ref: '#/components/schemas/AppFanCurve'
          description: Fan curve
          headers:
            ETag:
              examples:
                etag-example:
                  value: b5395865858689cfcb718a8fbf84e128771f6975
              explode: false
              schema:
                type: string
              style: simple
        "400":
          description: The request failed due to a malformed request body
      summary: Adds new fan curve
      x-openapi-router-controller: app.web.api.controllers.default_controller
    summary: Enables user to view all fan curves or to add a new one (up to 10 max)
  /app/fanCurves/{id}:
    delete:
      description: Deletes fan curve whose id correspond to specified `id`
      operationId: app_fan_curves_id_delete
      parameters:
      - description: Id of the fan curve (generated, i.e., surrogate key)
        examples:
          fanCurveId:
            value: 4731ab6a-433b-11ec-8321-c3a754deb306
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: Empty body
        "404":
          description: Requested fan curve w/ `id` doesn't exist
        "423":
          description: Fan curve corresponding to `id` cannot be deleted since fan
            curve is currently in use
      summary: Deletes fan curve whose id correspond to specified "id"
      x-openapi-router-controller: app.web.api.controllers.default_controller
    description: Singleton resource for fan curves
    get:
      description: Returns requested fan curve whose id corresponds to specified `id`
      operationId: app_fan_curves_id_get
      parameters:
      - description: Id of the fan curve (generated, i.e., surrogate key)
        examples:
          fanCurveId:
            value: 4731ab6a-433b-11ec-8321-c3a754deb306
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              examples:
                app-fanCurves_singleton:
                  value:
                    id: 916CD0EB-A755-4663-8410-461431039F74
                    name: Quiet
                    fanCurveSeries:
                    - tempInCels: 30
                      fanDCInPerc: 40
                    - tempInCels: 35
                      fanDCInPerc: 50
              schema:
                $ref: '#/components/schemas/AppFanCurve'
          description: Fan curve
          headers:
            ETag:
              examples:
                etag-example:
                  value: b5395865858689cfcb718a8fbf84e128771f6975
              explode: false
              schema:
                type: string
              style: simple
        "404":
          description: Requested fan curve w/ `id` doesn't exist
      summary: Returns requested fan curve whose id corresponds to specified "id"
      x-openapi-router-controller: app.web.api.controllers.default_controller
    put:
      description: Updates requested fan curve whose id corresponds to specified `id`
      operationId: app_fan_curves_id_put
      parameters:
      - description: Id of the fan curve (generated, i.e., surrogate key)
        examples:
          fanCurveId:
            value: 4731ab6a-433b-11ec-8321-c3a754deb306
        explode: false
        in: path
        name: id
        required: true
        schema:
          type: string
        style: simple
      - description: Current hash of to be updated fan curve -> used for optimistic
          locking
        examples:
          If-Match-example:
            value: b5395865858689cfcb718a8fbf84e128771f6975
        explode: false
        in: header
        name: If-Match
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            examples:
              app-fanCurvesBase_singleton:
                value:
                  name: Quiet
                  fanCurveSeries:
                  - tempInCels: 30
                    fanDCInPerc: 40
                  - tempInCels: 35
                    fanDCInPerc: 50
            schema:
              $ref: '#/components/schemas/AppFanCurveBase'
        description: Must contain the the updated fan curve corresponding to `id`
        required: true
      responses:
        "200":
          content:
            application/json:
              examples:
                app-fanCurves_singleton:
                  value:
                    id: 916CD0EB-A755-4663-8410-461431039F74
                    name: Quiet
                    fanCurveSeries:
                    - tempInCels: 30
                      fanDCInPerc: 40
                    - tempInCels: 35
                      fanDCInPerc: 50
              schema:
                $ref: '#/components/schemas/AppFanCurve'
          description: Fan curve
          headers:
            ETag:
              examples:
                etag-example:
                  value: b5395865858689cfcb718a8fbf84e128771f6975
              explode: false
              schema:
                type: string
              style: simple
        "400":
          description: The request failed due to a malformed request body
        "404":
          description: Requested fan curve w/ `id` doesn't exist
      summary: Updates requested fan curve whose id corresponds to specified "id"
      x-openapi-router-controller: app.web.api.controllers.default_controller
    summary: Allows user to view, update or delete fan curve corresponding to specified
      "id"
...
  /system/topTenProcesses/{nr}:
    description: 'Singleton resource for top 10 processes (dynamically aggregated;
      NOTE: MIGHT NOT WORK IN DOCKER)'
    get:
      description: Returns requested process whose current CPU utilization corresponds
        to specified `nr` in ranking
      operationId: system_top_ten_processes_nr_get
      parameters:
      - description: Corresponds to process w/ `nr` highest CPU utilization at current
          moment
        examples:
          topTenProcessNr:
            value: "2"
        explode: false
        in: path
        name: nr
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              examples:
                topTenProcesses_singleton:
                  value:
                    pid: 5424
                    ppid: 5434
                    name: "yes"
                    cpuUtilInPerc: 100
              schema:
                $ref: '#/components/schemas/SystemProcess'
          description: A process within the top ten processes
        "404":
          description: Requested process corresponding to `nr` doesn't exist
      summary: Returns requested process whose current CPU utilization corresponds
        to specified nr in ranking
      x-openapi-router-controller: app.web.api.controllers.default_controller
    summary: Allows user to view process w/ nr highest CPU utilization at current
      moment
components:
...
  schemas:
...
    SystemProcess:
      description: Represents process on host system (or in container if run in Docker
        w/ limited privileges)
      example:
        pid: 5424
        ppid: 5434
        name: "yes"
        cpuUtilInPerc: 100
      properties:
        pid:
          description: Pid of the process
          example: 5424
          readOnly: true
          type: integer
        name:
          description: Name of the process (or rather executable)
          example: "yes"
          readOnly: true
          type: string
        cpuUtilInPerc:
          description: CPU utilization of the process in percent
          example: 100.0
          maximum: 100
          readOnly: true
          type: number
        ppid:
          description: Pid of the parent process
          example: 5434
          readOnly: true
          type: integer
      required:
      - cpuUtilInPerc
      - name
      - pid
      - ppid
      title: Root Type for Process
      type: object
    AppFanCurveSeriesPoint:
      description: Point in fan curve series consisting of temperature in Celsius
        and fan speed in percent
      example:
        tempInCels: 30
        fanDCInPerc: 40
      properties:
        tempInCels:
          description: The temperature in Celsius for specified `fanDCInPerc`
          example: 40
          maximum: 130
          minimum: -10
          type: integer
        fanDCInPerc:
          description: The fan speed for specified `tempInCels` in percent
          example: 30
          maximum: 100
          minimum: 0
          type: integer
      required:
      - fanDCInPerc
      - tempInCels
      title: Root Type for FanCurveSeriesPoint
      type: object
...
    AppFanCurveBase:
      description: Fan curve consisting of fan curve series data and name (used for
        `POST`).
      example:
        name: Quiet
        fanCurveSeries:
        - tempInCels: 30
          fanDCInPerc: 40
        - tempInCels: 35
          fanDCInPerc: 50
      properties:
        name:
          description: Name of the fan curve
          example: Quiet
          maxLength: 255
          minLength: 0
          type: string
        fanCurveSeries:
          description: Array comprised of all fan series points
          example: |-
            [
                {
                    "tempInCels": 30,
                    "fanDCInPerc": 40
                },
                {
                    "tempInCels": 35,
                    "fanDCInPerc": 50
                }
            ]
          items:
            $ref: '#/components/schemas/AppFanCurveSeriesPoint'
          maxItems: 10
          minItems: 1
          type: array
          uniqueItems: true
      required:
      - fanCurveSeries
      - name
      title: AppFanCurveBase
      type: object
    AppFanCurve:
      allOf:
      - $ref: '#/components/schemas/AppFanCurve_allOf'
      - $ref: '#/components/schemas/AppFanCurveBase'
      description: Fan curve consisting of id, fan curve series data and name (used
        for `GET`).
      example:
        id: 916CD0EB-A755-4663-8410-461431039F74
        name: Quiet
        fanCurveSeries:
        - tempInCels: 30
          fanDCInPerc: 40
        - tempInCels: 35
          fanDCInPerc: 50
      title: AppFanCurve
      type: object
    AppConfig:
      description: Configuration flags pertaining the app itself and the pwm
      example:
        fanOn: true
        loggingEnabled: true
        loggingLevel: WARN
        DCUpdateIntervalInSec: 3
        selectedFanCurve:
          id: 916CD0EB-A755-4663-8410-461431039F74
          name: Quiet
          fanCurveSeries:
          - tempInCels: 30
            fanDCInPerc: 40
          - tempInCels: 35
            fanDCInPerc: 50
        pwmGpioPin: 12
        pwmInvertSignal: true
        pwmMinDCInPerc: 20
        pwmMaxDCInPerc: 95
      properties:
        fanOn:
          description: Whether fan is turned on (kill switch)
          example: true
          type: boolean
        loggingLevel:
          description: Current log level
          enum:
          - DEBUG
          - INFO
          - WARN
          - ERROR
          - CRITICAL
          example: WARN
          type: string
        DCUpdateIntervalInSec:
          description: How frequently the dutycycle shall be updated in seconds
          example: 3
          format: int32
          maximum: 10
          minimum: 1
          type: integer
        loggingEnabled:
          description: Whether logging is enabled
          example: true
          type: boolean
        selectedFanCurve:
          $ref: '#/components/schemas/AppFanCurve'
        pwmGpioPin:
          description: GPIO pin used for PWM (which applies the desired dutycycle
            for the fan)
          example: 12
          format: int32
          type: integer
        pwmInvertSignal:
          description: Whether PWM signal shall be inverted (may be required based
            on transitor type (PNP / NPN)
          example: true
          type: boolean
        pwmMaxDCInPerc:
          description: "Limits max duty cycle, i.e., the max fan speed in percent.\
            \ \nMay be desirable for some fans which only operate reliably within\
            \ a specific range of fan speeds.\nNOTE: This effectively overwrittes\
            \ the applied fan curve."
          example: 95
          format: int32
          maximum: 100
          minimum: 1
          type: integer
        pwmMinDCInPerc:
          description: "Limits min duty cycle, i.e., the min fan speed in percent.\
            \ \nMay be desirable for some fans which only operate reliably within\
            \ a specific range of fan speeds.\nNOTE: This effectively overwrittes\
            \ the applied fan curve."
          example: 20
          format: int32
          maximum: 100
          minimum: 1
          type: integer
      required:
      - DCUpdateIntervalInSec
      - fanOn
      - loggingEnabled
      - loggingLevel
      - pwmGpioPin
      - pwmInvertSignal
      - pwmMaxDCInPerc
      - pwmMinDCInPerc
      - selectedFanCurve
      title: AppConfig
      type: object
    AppFanCurve_allOf:
      properties:
        id:
          description: Id of fan curve
          example: 916CD0EB-A755-4663-8410-461431039F74
          maxLength: 255
          title: id
          type: string
      required:
      - id
      title: AppFanCurve_allOf
      type: object
  securitySchemes:
    api-security:
      description: Validated by webapp
      scheme: basic
      type: http
      x-basicInfoFunc: web.api.controllers.security_controller_.info_from_api_security